memperbaiki error dta stok masuk

+++ app/Database/Migrations/2024-04-27-145114_Satuan.php
@@ -10,7 +10,7 @@ class Satuan extends Migration
     {
         $this->forge->addField([
             'id_satuan' => [
-                'type'           => 'BIGINT',
+                'type'           => 'INT',
                 'constraint'     => 11,
                 'unsigned'       => true,
                 'auto_increment' => true,
